refactor(chart): refactor pie chart
